+++ spec/ig_markets/cli/commands/transactions_command_spec.rb
@@ -4,13 +4,13 @@ describe IGMarkets::CLI::Main, :cli_command do
   end
 
   before do
-    allow(Date).to receive(:today).and_return(Date.new(2016, 1, 5))
+    allow(Time).to receive(:now).and_return(Time.new(2016, 1, 5))
   end
 
   it 'prints transactions from a recent number of days' do
     transactions = [build(:transaction)]
 
-    expect(dealing_platform.account).to receive(:transactions).with(from: Date.new(2016, 1, 2)).and_return(transactions)
+    expect(dealing_platform.account).to receive(:transactions).with(from: Time.new(2016, 1, 2)).and_return(transactions)
 
     expect { cli(days: 3, interest: true, sort_by: 'date').transactions }.to output(<<-END
 #{IGMarkets::CLI::TransactionsTable.new transactions}
@@ -21,14 +21,14 @@ END
                                                                                    ).to_stdout
   end
 
-  it 'prints transactions from a number of days and a start date' do
-    from = Date.new 2015, 2, 15
-    to = Date.new 2015, 2, 18
+  it 'prints transactions from a number of days and a start time' do
+    from = Time.new 2015, 2, 15, 6, 0, 0
+    to = Time.new 2015, 2, 18, 6, 0, 0
 
     expect(dealing_platform.account).to receive(:transactions).with(from: from, to: to).and_return([])
 
     expect do
-      cli(days: 3, from: '2015-02-15', interest: true).transactions
+      cli(days: 3, from: '2015-02-15T06:00:00', interest: true).transactions
     end.to output("#{IGMarkets::CLI::TransactionsTable.new []}\n").to_stdout
   end
 
@@ -39,7 +39,7 @@ END
       build(:transaction, instrument_name: 'Test 456', profit_and_loss: 'US1.00')
     ]
 
-    expect(dealing_platform.account).to receive(:transactions).with(from: Date.new(2016, 1, 2)).and_return(transactions)
+    expect(dealing_platform.account).to receive(:transactions).with(from: Time.new(2016, 1, 2)).and_return(transactions)
 
     expect { cli(days: 3, instrument: 'TEST', interest: false, sort_by: 'date').transactions }.to output(<<-END
 #{IGMarkets::CLI::TransactionsTable.new [transactions[2], transactions[1]]}